Implement add team with routing

+++ src/app/teams/teams.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Renderer2, ElementRef, ViewChild } from '@angular/core';
 import {TeamService} from '../team.service';
 import { Team } from './team';
+import {Router, NavigationExtras} from "@angular/router";
 
 @Component({
   selector: 'app-teams',
@@ -13,7 +14,23 @@ export class TeamsComponent implements OnInit {
  
   selectedTeam : Team;
 
-  constructor(private teamService : TeamService) { }
+  @ViewChild('teams', {static: false}) list: ElementRef;
+
+  constructor(private teamService : TeamService,
+              private renderer: Renderer2,
+              private router: Router) {
+    this.renderer.listen('window', 'click',(e:Event)=>{
+      console.log("teams: got click event");
+      let elementId: string = (event.target as Element).id;
+      let className: string = (event.target as Element).className; 
+      console.log("teams: elementId="+elementId);
+      console.log("teams: className="+className);
+      if (className !== "selected")
+      {
+        this.selectedTeam=null;
+      }
+    });
+  }
 
   ngOnInit() {
     this.getTeams();
@@ -31,6 +48,29 @@ export class TeamsComponent implements OnInit {
 
   onDelete() {
     console.log("teams onDelete selectedTeam:"+this.selectedTeam.name);
-    this.teamService.deleteTeam(this.selectedTeam);
+    if(confirm("Are you sure you want to delete")) {
+      this.teamService.deleteTeam(this.selectedTeam);
+    }
   }
+
+  onAdd() {
+    console.log("teams onAdd");
+    this.router.navigateByUrl('/addteam');
+  }
+
+  onEdit() {
+    console.log("teams onEdit");
+    // this.router.navigateByUrl('editplayer');
+    let navigationExtras: NavigationExtras = {
+      queryParams: {
+          "id": this.selectedTeam.id,
+          "name": this.selectedTeam.name,
+          "record": this.selectedTeam.record,
+          "league": this.selectedTeam.league,
+          "players": this.selectedTeam.players
+      }
+    };
+    this.router.navigate(["editteam"], navigationExtras);
+  }
+
 }
